+++ AI/src/search/mcts/nodes/PNMCTSNode.java
@@ -10,7 +10,7 @@ import search.mcts.MCTS;
  * 
  * @author Dennis Soemers
  */
-public final class PNMCTSNode extends DeterministicNode
+public final class PNMCTSNode extends IPNMCTSNode
 {
 	
 	//-------------------------------------------------------------------------
@@ -62,12 +62,6 @@ public final class PNMCTSNode extends DeterministicNode
 	/** The value (in terms of proven/disproven/dont know) for this node */
 	protected PNMCTSNodeValues proofValue;
 	
-	/** Are the cached PNS-based terms of childrens' selection scores outdated? */
-	protected boolean childSelectionScoresDirty = false;
-	
-	/** Cached PNS-based terms for selection scores for all our children (including unexpanded ones) */
-	protected final double[] childrenPNSSelectionTerms;
-	
 	//-------------------------------------------------------------------------
     
     /**
@@ -111,8 +105,6 @@ public final class PNMCTSNode extends DeterministicNode
     	
     	evaluate();
         setProofAndDisproofNumbers();
-        
-        childrenPNSSelectionTerms = new double[numLegalMoves()];
     }
     
     //-------------------------------------------------------------------------
@@ -140,14 +132,7 @@ public final class PNMCTSNode extends DeterministicNode
         }
     }
     
-    /**
-     * Sets the proof and disproof values of the current node as it is done for 
-     * PNS in L. V. Allis' "Searching for Solutions in Games and Artificial 
-     * Intelligence". Set differently depending on if the node has children yet.
-     *
-     * @return Returns true if something was changed and false if not. 
-     * Used to improve PN-MCTS speed.
-     */
+    @Override
     public boolean setProofAndDisproofNumbers() 
     {
         if (legalMoves.length > 0) 
@@ -264,30 +249,11 @@ public final class PNMCTSNode extends DeterministicNode
     }
     
     /**
-     * @return Do our childrens' PNS-based selection terms need updating?
-     */
-    public boolean childSelectionScoresDirty()
-    {
-    	return childSelectionScoresDirty;
-    }
-    
-    /**
      * @return What type of node are we? (OR / AND)
      */
     public PNMCTSNodeTypes nodeType()
     {
     	return type;
-    }
-    
-    /**
-     * Store a flag saying whether the cached PNS-based terms of our childrens'
-     * selection scores are (potentially) outdated.
-     * 
-     * @param newFlag
-     */
-    public void setSelectionScoresDirtyFlag(final boolean newFlag)
-    {
-    	childSelectionScoresDirty = newFlag;
     }
     
     /**
@@ -304,14 +270,6 @@ public final class PNMCTSNode extends DeterministicNode
     public double disproofNumber()
     {
     	return disproofNumber;
-    }
-    
-    /**
-     * @return Array of PNS-based terms for selection strategy for all of our children.
-     */
-    public double[] childrenPNSSelectionTerms()
-    {
-    	return childrenPNSSelectionTerms;
     }
     
     //-------------------------------------------------------------------------
